issue, comment: moved out from jira package
